add styling to demo

+++ src/components/modal-content/DeleteOrganization.tsx
@@ -1,8 +1,9 @@
-import React, { useState } from 'react'
+import React, { useContext, useState } from 'react'
 import { Button, Message } from 'rsuite';
 import { toast } from 'react-toastify';
 
 import styles from "./deleteOrganization.module.scss";
+import { AuthContext } from '@/providers/AuthProvider';
 
 interface Props {
     onSuccess: () => void
@@ -12,18 +13,27 @@ interface Props {
 
 export const DeleteOrganization = (props: Props) => {
     const { organization, onSuccess, onCancel } = props;
-    
+
     const [isLoading, setIsLoading] = useState(false);
 
+    const { isDemo } = useContext(AuthContext);
+
     const handleDelete = async () => {
-        setIsLoading(true);
-        await fetch('/api/organizations', {
-            method: 'DELETE',
-            headers: {
-                'Content-Type': 'application/json',
-            },
-            body: JSON.stringify({ id: organization.id }),
-        }).finally(() => setIsLoading(false));
+
+        if (isDemo) {
+            const demoOrganizations = JSON.parse(localStorage.getItem("demoOrganizations") || "[]");
+            const remainingOrganizations = demoOrganizations.filter((demoOrganization: any) => demoOrganization.id !== organization.id);
+            localStorage.setItem("demoOrganizations", JSON.stringify(remainingOrganizations));
+        } else {
+            setIsLoading(true);
+            await fetch('/api/organizations', {
+                method: 'DELETE',
+                headers: {
+                    'Content-Type': 'application/json',
+                },
+                body: JSON.stringify({ id: organization.id }),
+            }).finally(() => setIsLoading(false));
+        }
         toast.success("Removed!");
         onSuccess();
     }
